less *special* block stacking
basically just refactored the stacking system to loop through relative block faces and see if those can be stacked. 

the old system was stupid, causing frequent floating holograms and block loses. this is mainly due to the fact that it relied on the BlockPlaceEvent#getBlockAgainst method, which would fire for canceled block places and was especially easy to break with higher pings.